+++ src/shinestacker/config/defaults.py
@@ -19,7 +19,8 @@ DEFAULTS = {
     'brush_flow': 100,
     'temp_folder_path': '',
     'raw_formats': {
-        'auto_bright': True
+        'auto_bright': True,
+        'auto_bright_thr': 1.0,   # %
     },
     'sequential_task': {
         'max_threads': 8,